Don't try to drop databases etc with checkstyle CI

+++ config/ci/init.sh
@@ -31,6 +31,13 @@ export KEYSTORE=${PWD}/keystore.jks
 
 echo "#---> Running $TEST_SUITE tests"
 
+if [ "$TEST_SUITE" = "checkstyle" ]; then
+    # TODO: It's pretty pointless to run this script with this suite
+    # Better to have custom scripts with shared common functions
+    # and fewer conditionals
+    exit 0 # nothing to do
+fi
+
 sudo -E -u postgres dropdb -h "$PSQL_HOST" --if-exists flatmodetest
 sudo -E -u postgres dropdb -h "$PSQL_HOST" --if-exists fulldatatest
 sudo -E -u postgres dropdb -h "$PSQL_HOST" --if-exists notxmltest
@@ -46,48 +53,44 @@ sudo -E -u postgres dropuser -h "$PSQL_HOST" --if-exists test
 sudo -E -u postgres createuser -h "$PSQL_HOST" test
 sudo -E -u postgres psql -h "$PSQL_HOST" -c "alter user test with encrypted password 'test';"
 
-if [ "$TEST_SUITE" = "checkstyle" ]; then
-    exit 0 # nothing to do
-else
-    # Set up properties
-    source "${WORKSPACE_DIR}"/config/create-ci-properties-files.sh
+# Set up properties
+source "${WORKSPACE_DIR}"/config/create-ci-properties-files.sh
 
-    echo '#---> Installing python requirements'
-    # Install lib requirements
-    ${PYTHON} -m pip install -r "${WORKSPACE_DIR}"/config/lib/requirements.txt
+echo '#---> Installing python requirements'
+# Install lib requirements
+${PYTHON} -m pip install -r "${WORKSPACE_DIR}"/config/lib/requirements.txt
 
-    if [[ "$TEST_SUITE" = "bio" ]]; then
-        (cd "${WORKSPACE_DIR}"/plugin && ./gradlew install --info)
-        (cd "${WORKSPACE_DIR}"/intermine && ./gradlew intermine-webapp:war --info)
-    fi
+if [[ "$TEST_SUITE" = "bio" ]]; then
+    (cd "${WORKSPACE_DIR}"/plugin && ./gradlew install --info)
+    (cd "${WORKSPACE_DIR}"/intermine && ./gradlew intermine-webapp:war --info)
+fi
 
-    if [[ "$TEST_SUITE" = "ws" ]]; then
+if [[ "$TEST_SUITE" = "ws" ]]; then
 
-        # install everything first. we don't want to test what's in maven
-        (cd "${WORKSPACE_DIR}"/plugin && ./gradlew install)
-        (cd "${WORKSPACE_DIR}"/intermine && ./gradlew install)
-        (cd "${WORKSPACE_DIR}"/bio && ./gradlew install)
-        (cd "${WORKSPACE_DIR}"/bio/sources && ./gradlew install)
-        (cd "${WORKSPACE_DIR}"/bio/postprocess && ./gradlew install)
+    # install everything first. we don't want to test what's in maven
+    (cd "${WORKSPACE_DIR}"/plugin && ./gradlew install)
+    (cd "${WORKSPACE_DIR}"/intermine && ./gradlew install)
+    (cd "${WORKSPACE_DIR}"/bio && ./gradlew install)
+    (cd "${WORKSPACE_DIR}"/bio/sources && ./gradlew install)
+    (cd "${WORKSPACE_DIR}"/bio/postprocess && ./gradlew install)
 
-        # set up database for testing
-        (cd "${WORKSPACE_DIR}"/intermine && ./gradlew createUnitTestDatabases)
+    # set up database for testing
+    (cd "${WORKSPACE_DIR}"/intermine && ./gradlew createUnitTestDatabases)
 
-        # We will need a fully operational web-application
-        echo '#---> Building and releasing web application to test against'
-        (cd "${WORKSPACE_DIR}"/testmine && ./setup.sh "${WORKSPACE_DIR}")
+    # We will need a fully operational web-application
+    echo '#---> Building and releasing web application to test against'
+    (cd "${WORKSPACE_DIR}"/testmine && ./setup.sh "${WORKSPACE_DIR}")
 
-        # Warm up the keyword search by requesting results, but ignoring the results
-        $GET "$TESTMODEL_URL/service/search" > /dev/null
-        # Start any list upgrades
-        $GET "$TESTMODEL_URL/service/lists?token=test-user-token" > /dev/null
+    # Warm up the keyword search by requesting results, but ignoring the results
+    $GET "$TESTMODEL_URL/service/search" > /dev/null
+    # Start any list upgrades
+    $GET "$TESTMODEL_URL/service/lists?token=test-user-token" > /dev/null
 
-        cd "${WORKSPACE_DIR}"
-        if [[ "$CLIENT" = "JS" ]]; then
-            # We need the imjs code to exercise the webservices
-            $GIT_GET https://github.com/intermine/imjs.git client-JS
-        elif [[ "$CLIENT" = "PY" ]]; then
-            $GIT_GET -b dev https://github.com/intermine/intermine-ws-python client-PY
-        fi
+    cd "${WORKSPACE_DIR}"
+    if [[ "$CLIENT" = "JS" ]]; then
+        # We need the imjs code to exercise the webservices
+        $GIT_GET https://github.com/intermine/imjs.git client-JS
+    elif [[ "$CLIENT" = "PY" ]]; then
+        $GIT_GET -b dev https://github.com/intermine/intermine-ws-python client-PY
     fi
 fi
